pyproject.toml
[project]
    name = "llm-p"
    version = "0.1.0"
    description = "FastAPI service with JWT auth, SQLite, and OpenRouter LLM proxy"
    requires-python = ">=3.11"
    dependencies = [
      "fastapi>=0.112.0",
      "uvicorn[standard]>=0.30.0",
      "sqlalchemy>=2.0.30",
      "aiosqlite>=0.20.0",
      "pydantic[email]>=2.7.0",
      "pydantic-settings>=2.3.0",
      "python-jose[cryptography]>=3.3.0",
      "passlib[bcrypt]>=1.7.4",
      "httpx>=0.27.0",
      "python-multipart>=0.0.9",
      "greenlet>=3.3.1",
      "bcrypt==3.2.0",
      "ruff>=0.14.14",
    ]
    
    [dependency-groups]
    dev = []